[FEAT]: gradle 추가한 프로젝트로 변경
