+++ project/stampee/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Wed Jul 24 10:26:52 KST 2024
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.5-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
