--- project/stampee/src/domain/Cafe.java
@@ -1,7 +1,0 @@
-package domain;
-
-public class Cafe {
-    private long cafeId;
-    private String name;
-    private String address;
-}
