--- project/stampee/src/domain/Member.java
@@ -1,9 +1,0 @@
-package domain;
-
-public class Member {
-  private long memberId;
-  private String password;
-  private String email;
-  private String phoneNumber;
-
-}
