--- project/stampee/src/domain/Signature.java
@@ -1,11 +1,0 @@
-package domain;
-
-import java.util.List;
-
-public class Signature {
-
-  private long menuId;
-  private String menuName;
-  private List<Cafe> cafes;
-  
-}
