+++ project/stampee/src/main/java/domain/Cafe.java
@@ -1,0 +1,7 @@
+package domain;
+
+public class Cafe {
+    private long cafeId;
+    private String name;
+    private String address;
+}
