+++ project/stampee/src/main/java/domain/Member.java
@@ -1,0 +1,9 @@
+package domain;
+
+public class Member {
+	private long memberId;
+	private String password;
+	private String email;
+	private String phoneNumber;
+
+}
