+++ project/stampee/src/main/java/domain/Signature.java
@@ -1,0 +1,11 @@
+package domain;
+
+import java.util.List;
+
+public class Signature {
+
+	private long menuId;
+	private String menuName;
+	private List<Cafe> cafes;
+
+}
